requirements.txt
autopep8>=1.5.5
certifi>=2020.12.5
cffi>=1.14.4
click>=7.1.2
constantly>=15.1.0
cryptography>=3.3.1
cssselect>=1.1.0
dateparser>=1.0.0
python-dateutil>=2.8.1
mongoengine>=0.23.0
numpy>=1.20.0
pandas>=1.2.1
python-dotenv>=0.15.0
PyMongo>=3.11.2
pytest>=6.2.3
requests>=2.25.1
word2number>=1.1
spacy>=3.0.0
tqdm
marshmallow>=3
flask
flask-cors
flask-restful
flask-migrate
Flask-PyMongo
flask-sqlalchemy
flask-marshmallow
flask-jwt-extended
marshmallow-sqlalchemy
python-dotenv
passlib
apispec[yaml]
apispec-webframeworks
tox
celery[redis]>=5.0.0
gunicorn
